requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.15
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
datasets==4.1.0
dill==0.4.0
distro==1.9.0
filelock==3.19.1
frozenlist==1.5.0
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.0
google-auth-httplib2==0.2.0
google-genai==1.39.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.14.0
hf-xet==1.1.10
httpcore==1.0.7
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.34.6
idna==3.10
jiter==0.9.0
multidict==6.3.0
multiprocess==0.70.16
numpy==2.2.4
nums-from-string==0.1.2
openai==1.77.0
packaging==25.0
pandas==2.2.3
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.1
pydantic_core==2.33.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.18.3
